requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-crispy-forms==1.13.0
pillow==10.4.0
sqlparse==0.5.1
typing-extensions==4.12.2
